+++ crates/rhdl-fpga/src/dsp/mod.rs
@@ -1,3 +1,4 @@
 //! DSP Related Cores
+pub mod iq;
 pub mod lerp;
 pub mod nco;